refactor(binary-index): unify the three FactKeyV3 definitions behind one canonical type
The V3 fact-identity tuple (s_id, p_id, o_type, o_key, o_i) was defined
three times with identical fields but divergent derives: the pub one in
read/types.rs (cursor merge, resolve_overlay_ops), a private copy in
read/replay.rs (time-travel replay membership), and a third in the
indexer's novelty_merge (build-time history dedup). The OI_NONE o_i
sentinel was likewise duplicated.

Promote read/types.rs::FactKeyV3 to the single definition: lift the
from_batch / from_hist constructors and a pub OI_NONE onto it, and have
replay and the indexer import it. The canonical derive set is a superset
of both removed copies, so this is behavior-preserving; the only
adjustment is dropping a now-redundant clone (clone_on_copy) in replay.

Also documents the two legitimate overlay fact-identity domains on the
canonical type: the encoded V3 key here vs the row-world Flake tuple
(s, p, o, dt, m) used by the raw-flake post-pass, which is required
because FactKeyV3 cannot express a language tag (#1273).

Phase 1B of the overlay-contract refactor (audit roadmap).
